refactor: Rename files for consistent naming conventions
Renamed i18n_dev.ts to i18n-dev.ts and i18n_prod.ts to i18n-prod.ts. Updated import statements to reflect the new file names. This change ensures consistency and improves readability in the codebase.
