+++ packages/i18n/lib/i18n-prod.ts
@@ -4,4 +4,4 @@ export function t(key: MessageKey, substitutions?: string | string[]) {
   return chrome.i18n.getMessage(key, substitutions);
 }
 
-t.devLocale = '';
+t.devLocale = ''; // for type consistency with i18n-dev.ts
